Starvation now works, as well as natural pipe color
All of rivulets mechanics seem to be fixed now that their speed and lungs are scaled. They probably actually have better lungs because slugpups get a lung increase. The jump and backflip seem to not have been broken in the first place, the way they get scaled with pups jumping ability makes those traits seem similar to a regular slugcat rather than a pup. Backflip might not have the height it's supposed to, however, I'll need to actually check the game code for that to confirm.

Incorrect number of food pips is displayed on the menu. That seems to be the last visual bug that I need to take care of.