Figured out how to set up stomach controls
Downside is that any slugcat that starts with an item in their stomach or has one when the option is turned on will have that item trapped inside until they turn it off. I experimented with taking it out on the first frame but slugcats that start with an item in their stomach also start with an opening cutscene, which lead to all kinds of errors. This is the simpler solution and players simply need to take a cycle to disable the option if they want to retrieve the item.