Players in multiplayer now have different stats
FIRST SUCCESSFUL ILHOOK LETS GOOOOOOOOO

only seconds players stats seem completely correct currently, need to hunt down what messes up the first player

This update messes up most stats and completely turns the food option off, on top of not working completely correctly. There will need to be quite a few updates and some testing before this can even be updated on steam. It also only works with two players currently. The main purpose of this was to see in a simpler context what needs to be changed and altered and finding the edge cases before I expand it to work it 4 players (and maybe many more if I want to be compatible with the upcoming 16 player mod)