Made it so arena doesn't crash with new stat assignments
This was surprisingly hard to do, I expected it to be at least as straightforward as last time but apparently arena works so differently that a ton of if statements are needed. I cleaned up as well as I could.

Doesn't work as intended yet. Only one of the players stats are used again and for some reason all the players change classes and controls every time they enter the arena even though the ui still assigns to the correct class instead of controller (i.e. if player 1 (spearmaster) suddenly takes on player 2s (Rivulets) color scheme and controller, then spearmaster wins, Player 1 still gets the point even though it was essentially player 2s character at that point) Then both player work correctly every other match. This is an incredibly weird glitch that I have no idea how was caused as of right now.