All Stat and Player assignments are now correct
This is true for Story and Competitive Arena mode (while jolly coop is enabled), for the 4 players the game innately supports.

The game immediately crashes in sandbox and challenge mode, so with little testing, I assume the number of starting Increments needs to be changed to get them to work. I have not tested expedition at all yet, though I expect the fix to be just as simple in that case.

I believe I've gotten most of the hard parts for the next update out of the way today. The update after that, when I try to let some players play as adults, will likely be a different story where I need to do a lot more refactoring.