Fixed sandbox mode
Sandbox mode actually wasn't broken, a natural pup was spawning in my sandbox world and the new system didn't handle that correctly.

This update made it so that slugpups could spawn properly (at least in arena mode), guaranteed that players some player IDs can properly be skipped over in multiplayer, as well as fixed up quite a few mistakes I made in programming the new system thus far.

This is really just as painful to debug as the multiplayer implementation my first go-around was. There are too many moving parts with strange interactions depending on the game mode.